requirements.txt
cachetools
certifi
cffi
charset-normalizer
colorclass
cryptography
docopt
docx2txt
fpdf
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
httplib2
idna
imap-tools
langdetect
lxml
packaging
pdf2image
pdfminer.six
pdfplumber
pillow
pip-upgrader
plum-dispatch
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pycryptodome
PyMuPDF
pyparsing
pypdfium2
pytesseract
python-dateutil
python-docx
requests
rsa
rtf-converter
schedule
six
terminaltables
typing_extensions
uritemplate
urllib3
striprtf
pypdf
pyinstaller
google-api-python-client-stubs
pinecone
pinecone-plugin-assistant
azure-ai-formrecognizer>=3.3.0
azure-core>=1.28.0
google-cloud-translate>=3.15.0
PyPDF2>=3.0.0
paddleocr>=2.7.0
paddlepaddle>=3.2.0
img2pdf>=0.5.0